2392: move generic TextField outside of annotation
