+++ src/common/fields/TextField.js
@@ -6,8 +6,9 @@ import {
 import PropTypes from 'prop-types';
 import React, { useMemo } from 'react';
 
-import { useTranslate } from '../../i18n/I18NContext';
+import { useTranslate } from '../../app/js/i18n/I18NContext';
 
+// TextField component to use tanstack react form with material ui text field
 export function TextField({
     form,
     name,
